Refactor create/delete ad apis

+++ server/src/main/java/hu/personal/caradvertiser/service/AdvertisementService.java
@@ -8,6 +8,8 @@ import hu.personal.caradvertiser.model.FilterDto;
 import hu.personal.caradvertiser.model.FilterResultDto;
 import hu.personal.caradvertiser.repository.AdvertisementRepository;
 import lombok.RequiredArgsConstructor;
+import org.springframework.security.core.Authentication;
+import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.stereotype.Service;
 
 import javax.persistence.EntityNotFoundException;
@@ -27,6 +29,8 @@ public class AdvertisementService {
     }
 
     public AdRegisterResponseDto createAd(AdvertisementDto advertisementDto) {
+        Authentication authentication = SecurityContextHolder.getContext().getAuthentication();
+        advertisementDto.setUsername(authentication.getName());
         Advertisement savedAdvertisement = advertisementMapper.toEntity(advertisementDto);
         AdvertisementDto savedAdvertisementDto = advertisementMapper.toDto(advertisementRepository.save(savedAdvertisement));
         return new AdRegisterResponseDto().id(savedAdvertisementDto.getId());
@@ -34,6 +38,11 @@ public class AdvertisementService {
 
     public AdvertisementDto deleteAd(Long id) {
         Advertisement advertisement = advertisementRepository.findById(id).orElseThrow(EntityNotFoundException::new);
+        Authentication authentication = SecurityContextHolder.getContext().getAuthentication();
+        String currentUsername = authentication.getName();
+        if (!currentUsername.equals(advertisement.getUsername())) {
+            throw new IllegalArgumentException();
+        }
         advertisementRepository.delete(advertisement);
         return advertisementMapper.toDto(advertisement);
     }
